flux_lora_style_tool/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
Pillow>=9.0.0
numpy>=1.21.0
matplotlib>=3.5.0
opencv-python>=4.5.0
PyYAML>=6.0
tqdm>=4.64.0
scikit-image>=0.19.0
imageio>=2.19.0
